PA2 과제 오류 일부 수정 + 실행 가능한 파일 추가

+++ pa2-xv6-riscv/kernel/defs.h
@@ -60,7 +60,7 @@ void*           kalloc(void);
 void            kfree(void *);
 void            kinit(void);
 // kalloc.c에 정의된 함수
-uint64          freememinfo(void);
+int             freememinfo(void);
 
 // log.c
 void            initlog(int, struct superblock*);
